requirements.txt
flask
gunicorn
qdrant-client
mediapipe
tensorflow-cpu
opencv-python-headless
requests
numpy